feature: implement responsive layout
