+++ src/components/pageFooter.tsx
@@ -1,0 +1,9 @@
+export default function PageFooter() {
+  return (
+    <footer
+      className={'bg-green-400 p-2 laptop:col-start-2 laptop:row-start-4'}
+    >
+      <b>Hello from the footer...</b>
+    </footer>
+  )
+}
