+++ src/components/sidebar.tsx
@@ -1,0 +1,18 @@
+export default function Sidebar() {
+  return (
+    <section
+      className={
+        'bg-amber-400 p-5 prose max-w-none laptop:col-start-1 laptop:row-start-1 laptop:row-span-4'
+      }
+    >
+      <h1>Hi! I'm Connor</h1>
+      <p>This is the sidebar container</p>
+      <p>
+        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at
+        consequatur ea ex minus, rem reprehenderit velit. Commodi enim iste modi
+        voluptate? Accusantium asperiores, aut fugiat laudantium minima
+        obcaecati veritatis?
+      </p>
+    </section>
+  )
+}
