+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@
 import '../styles/globals.css'
 import type { AppType } from 'next/dist/shared/lib/utils'
 import { trpc } from '../utils/trpc'
-import Layout from '../components/layout/layout'
+import Layout from '../components/layout'
 
 const MyApp: AppType = ({ Component, pageProps }) => {
   return (
